requirements.txt
streamlit
pypdf2
google-generativeai